made the typing text it's own component
